Set `$attribution_locale` in generated muttrc
The French attribution from before capitalised the month name because I
have an EN locale by default. The month shouldn't be capitalised in
French. Set the `$attribution_locale` to match the language of the
attribution template string.

Hard-coding `en_US` on the English one as that's the locale I'm used to.